EASY-990 add streaming surrogate relation to ddm, EASY-996 (partial) upgrade ddm.xsd+examples
Moved OfflineDdmValidator to a test package and as the name was
misguided renamed it to SpecialValidator with javadoc to explain 'special'

Either the published schemas or snapshot versions (the last build)
of easy-schema are now installed in the target folder of easy-app/lib/ddm

Restored various ignored tests with an assume that checks availability
of external schemas.

Schema changes and validations make easy-schema and ddm
mutualy(!) dependent. Until easy-schema is merged and published
in the maven repository:
- pull easy-schema (if aplicable) and build it
- execute 'mvn -DSNAPSHOT_SCHEMA=true clean install'
  for easy-app/lib/ddm

Change-Id: Ia04280257856ea21249ae60ad3949166a47bd3b5